layers: Add GpuAssistedBase class
This is a common base class for DebugPrintf and GPUAV.
